pyproject.toml
[project]
name = "data-pipeline-vbplnew"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dlt[postgres]>=1.24.0",
    "environs>=14.6.0",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.0",
    "langchain-core>=1.2.20",
    "langchain-huggingface>=1.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-pinecone>=0.2.13",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "markdownify>=1.2.2",
    "pinecone>=7.3.0",
    "plotext>=5.3.2",
    "psycopg2-binary>=2.9.11",
    "scrapy>=2.14.2",
    "scrapy-user-agents>=0.1.1",
    "sentence-transformers>=5.3.0",
    "torch>=2.10.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
]
